backend/app/config.py
from dotenv import load_dotenv
import os

# Load environment variables from .env
load_dotenv()

BINANCE_WS_URL = os.getenv("BINANCE_WS_URL")

TWILIO_ACCOUNT_SID = os.getenv("TWILIO_ACCOUNT_SID")
TWILIO_AUTH_TOKEN = os.getenv("TWILIO_AUTH_TOKEN")
TWILIO_PHONE_NUMBER = os.getenv("TWILIO_PHONE_NUMBER")
USER_PHONE_NUMBER = os.getenv("USER_PHONE_NUMBER")
