backend/requirements.txt
fastapi
uvicorn
python-dotenv
twilio
websockets
httpx
